revert commit 'added interface for each service class'
